ml_training/requirements.txt
torch==2.1.0
torchvision==0.16.0
coremltools==7.1
Pillow==10.1.0
requests==2.31.0
python-dotenv==1.0.0
tqdm==4.66.1
numpy==1.24.3
